requirements.txt
streamlit>=1.31.0
groq>=0.4.0
textblob>=0.17.1
matplotlib>=3.8.0
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0

